+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+- WorkerFactoryInterface::create() receives a second argument, the worker name
+
 ## [0.1.0] - 2016-12-27
 Initial public release